==> 03_Analysis_europe.Rmd <==
---
title: "Covid Data and Restrictions - Regression Models - Europe"
output: github_document
always_allow_html: yes
---
```{r setup, echo=FALSE, message=FALSE, warning=FALSE} 
knitr::opts_chunk$set(warning = FALSE, message = FALSE) 
```

```{r, echo=FALSE, message=FALSE, warning=FALSE}
if (!require(pacman)) install.packages('pacman')
library(pacman)
p_load(tidyverse,ggpubr,ggrepel,ggtext,scales,viridis,ggthemes,psych,kableExtra,corrr,car,ggiraphExtra)
options(scipen=999)
options(warn=-1)
```

```{r,echo=FALSE, message=FALSE, warning=FALSE}
lally_countries <- c("Austria","Albania","Belgium","Bosnia and Herzegovina","Bulgaria","Croatia","Cyprus","Czechia","Denmark","Estonia","Finland","France","Germany","Greece","Hungary","Iceland","Ireland","Italy","Latvia","Lithuania","Luxembourg","Netherlands","Norway","Poland","Portugal","Romania","Serbia","Slovakia","Slovenia","Spain","Sweden","Switzerland","United Kingdom")

lally_countries_map <- c("Austria","Albania","Belgium","Bosnia and Herzegovina","Bulgaria","Croatia","Cyprus","Czech Republic","Denmark","Estonia","Finland","France","Germany","Greece","Hungary","Iceland","Ireland","Italy","Latvia","Lithuania","Luxembourg","Netherlands","Norway","Poland","Portugal","Romania","Serbia","Slovakia","Slovenia","Spain","Sweden","Switzerland","UK")


df_europe_20201230 <- readRDS("OxOw_Extraction_europe_20201230_large.rds") %>% filter(location %in% lally_countries)
 
df_europe_20211230 <- readRDS("OxOw_Extraction_europe_20211230_large.rds") %>% filter(location %in% lally_countries)

df_europe_20220520 <- readRDS("OxOw_Extraction_europe_20220520_large.rds") %>% filter(location %in% lally_countries)
```

```{r,echo=FALSE, message=FALSE, warning=FALSE}
cor_variables <- c("location","iso_code","date","continent","total_deaths_per_million","people_fully_vaccinated_per_hundred","population_density","median_age","gdp_per_capita","extreme_poverty","cardiovasc_death_rate","diabetes_prevalence","life_expectancy","human_development_index","latitude","sub_stringency","sub_stringency_cubed","first_death_date","date_first_death_days","stringency","obesity_prevalence","lockdowns","lockdowns_cubed")

df_europe_20220520 <- df_europe_20220520  %>% select(cor_variables)
df_europe_20211230 <- df_europe_20211230  %>% select(cor_variables)
df_europe_20201230 <- df_europe_20201230  %>% select(cor_variables)


corr_table <- function(df){
  date <- df$date[1]
  tble <- df[,unlist(lapply(df, is.numeric))] %>% correlate() %>% focus(total_deaths_per_million) %>% 
  rename(Variable = term) %>% arrange(total_deaths_per_million) %>%  
  kbl(caption = paste0("Correlation - Europe - Total deaths per million - ",date)) %>% 
  kable_classic(full_width = F)
  
  return(tble)
}


corr_plot <- function (df){
  date <- df$date[1]
  plot <- df[,unlist(lapply(df, is.numeric))] %>% correlate() %>% focus(total_deaths_per_million) %>% 
  mutate(term = factor(term, levels = term[order(total_deaths_per_million)])) %>%
  ggplot(aes(x = term, y = total_deaths_per_million)) +
    geom_bar(stat = "identity", fill = 'purple') +
    ylab("Total deaths per million (corr)") +
    xlab("") +
    ylim(c(-1,1)) +
    ggtitle(paste0("Correlation - Europe - Total deaths per million - ",date)) +
    theme_classic2() +
    theme(axis.text.x=element_text(angle=75,hjust=1,vjust=1))
  
  return(plot)
  
}

corr_matrix <- function(df){
  date <- df$date[1]
  corPlot(df[,unlist(lapply(df, is.numeric))],
        min.length = 20,
        scale = FALSE,
        diag = FALSE,
        upper = FALSE,
        alpha = 0.20,
        stars = TRUE,
        cex = 0.5,
        cex.axis = 0.55,
        xlas = 2,
        main = paste0("Correlations - Europe -  ",date),
        )
  
}
mlr_table_si <- function(df){
  
  capture.output(summary(lm(formula = total_deaths_per_million ~ stringency + population_density +
              date_first_death_days, data = df)), file=NULL,append=FALSE) %>% 
  as.tibble() %>% 
  rename(regression = value) %>% 
  kbl() %>% 
  kable_classic()
}

mlr_table_res <- function(df){
  
  capture.output(summary(lm(formula = total_deaths_per_million ~ sub_stringency + population_density +
              date_first_death_days, data = df)), file=NULL,append=FALSE) %>% 
  as.tibble() %>% 
  rename(regression = value) %>%
  kbl() %>% 
  kable_classic()
}

mlr_table_res_cubed <- function(df){
  
  capture.output(summary(lm(formula = total_deaths_per_million ~ sub_stringency_cubed + population_density +
              date_first_death_days, data = df)), file=NULL,append=FALSE) %>% 
  as.tibble() %>% 
  rename(regression = value) %>%
  kbl() %>% 
  kable_classic()
}

mlr_table_lock <- function(df){
  
  capture.output(summary(lm(formula = total_deaths_per_million ~ lockdowns + population_density +
              date_first_death_days, data = df)), file=NULL,append=FALSE) %>% 
  as.tibble() %>% 
  rename(regression = value) %>%
  kbl() %>% 
  kable_classic()
}

mlr_table_lock_cubed <- function(df){
  
  capture.output(summary(lm(formula = total_deaths_per_million ~ lockdowns_cubed + population_density +
              date_first_death_days, data = df)), file=NULL,append=FALSE) %>% 
  as.tibble() %>% 
  rename(regression = value) %>%
  kbl() %>% 
  kable_classic()
}

mlr_print <- function(model){
  capture.output(summary(model), file = NULL, append = FALSE) %>% 
  as.tibble() %>% 
  rename(regression = value) %>% 
  kbl() %>% 
  kable_classic()
}

  
mlr_vif <- function(model){
  vif_values <- vif(model)
  barplot(vif_values, main = "VIF Values", col = "steelblue", horiz=F, las=2,cex.names=.5)
  abline(h = 5, lwd = 3, lty = 2)}


create_model_full <- function(df){
  
  model <- lm(formula = total_deaths_per_million ~ median_age + people_fully_vaccinated_per_hundred + population_density + gdp_per_capita + 
                          extreme_poverty + cardiovasc_death_rate + diabetes_prevalence + life_expectancy + human_development_index + latitude + 
                          sub_stringency + date_first_death_days + obesity_prevalence, data = df)
  return (model)
}

create_model_stringency <- function(df){
  model <-  lm(formula = total_deaths_per_million ~ median_age + stringency + 
                people_fully_vaccinated_per_hundred,data = df)
  return(model)
} 

create_model_sub_stringency <- function(df){
  model <-  lm(formula = total_deaths_per_million ~ median_age + sub_stringency + 
                people_fully_vaccinated_per_hundred,data = df)
  return(model)
} 

create_model_sub_stringency_cubed <- function(df){
  model <-  lm(formula = total_deaths_per_million ~ median_age + sub_stringency_cubed + 
                people_fully_vaccinated_per_hundred,data = df)
  return(model)
} 

create_model_lockdowns <- function(df){
  model <-  lm(formula = total_deaths_per_million ~ median_age + lockdowns + 
                people_fully_vaccinated_per_hundred,data = df)
  return(model)
} 

create_model_lockdowns_cubed <- function(df){
  model <-  lm(formula = total_deaths_per_million ~ median_age + lockdowns_cubed + 
                people_fully_vaccinated_per_hundred,data = df)
  return(model)
} 
```

```{r,echo=FALSE, message=FALSE, warning=FALSE}
#map print function
country_map <- function(df,dt,fill_var,title,leg_label,country_list,include_title){
  
  if (include_title == 'yes') {Switch = T}
  else if((include_title == 'no')) {Switch = F}
  
  df_world <- df %>% filter(date == dt)

  map_data <- map_data("world", region = country_list) %>% 
    select(-c(order,subregion)) %>% 
    rename(location = region)
  
  map_data$location <- gsub("Czech Republic","Czechia",map_data$location)
  map_data$location <- gsub("USA", "United States", map_data$location)
  map_data$location <- gsub("UK", "United Kingdom", map_data$location)

  #merge dataframes based on country location
  map_df <- full_join(map_data,df_world, by = 'location') %>% 
  #select only specific columns
  select(location,sub_stringency,sub_stringency_cubed, total_deaths_per_million,
          stringency,lockdowns, lockdowns_cubed, group, long, lat)
  
  #summarise data to reduce to one country per row
  map_lab_data <- map_df %>%
  group_by(location) %>%
  summarise(long = mean(long), lat = mean(lat), group = max(group))

  map_lab_data  <- map_lab_data  %>% drop_na()
  map_df <- map_df %>% drop_na()
  
  
  
  map <- ggplot(map_df, aes(x = long, y = lat, group = group)) +
  geom_polygon(aes_string(fill = fill_var), colour = 'white') +
  scale_fill_viridis_c(option = "C", direction = -1) +
  labs(fill = leg_label) +
  {if(Switch)ggtitle(paste0(title, " Map ", date_input))} +
  geom_text_repel(data = map_lab_data, aes(label = location), max.overlaps = 7,
            size=3, min.segment.length = 2, colour = 'black',
            segment.alpha = 0,segment.color = 'grey50') +
  theme_map() +
  theme(    axis.title=element_blank(),
            axis.text=element_blank(),
            axis.ticks=element_blank(),
            plot.title = element_text(hjust = 0.5, size = 14, family="serif"),
            legend.text=element_text(size=10,family="serif"),
            legend.title=element_text(size=10,family="serif"),
            legend.position="right",
            legend.direction = "vertical")
   
  return(map)
}
```

```{r,echo=FALSE, message=FALSE, warning=FALSE}
#function to plot scatter plot 
country_scatter <- function(df_input, x_metric, x_label, title, equation = "no") {
  if(df_input %>% select(x_metric) %>% min(na.rm = T) >=1){
    min_x <- df_input %>% select(x_metric) %>% min(na.rm = T)
  }
  else{
    min_x <- df_input %>% select(x_metric) %>% min(na.rm = T)
  }
  
  max_x <- df_input %>% select(x_metric) %>% max(na.rm = T)
  max_y <- df_input %>% select('total_deaths_per_million') %>% max(na.rm = T)
  
  if (title == 'yes') {Switch = T}
  else {Switch = F}

  if (equation == 'ann') {ann = T}
  else {ann = F}

  date <- df_input %>% select(date) %>% filter(row_number()==1) %>% pull()
  plot <- df_input %>% 
ggplot(aes_string(x = x_metric, y = "total_deaths_per_million", label = "location")) +
        geom_point(data = df_input, aes(color = continent)) +
        geom_smooth(method = lm, se = F, colour = 'grey') +
        facet_wrap(~continent) +
        scale_size_continuous(range = c(2,10)) +
        scale_color_viridis(discrete = TRUE, option = 'C') +
        geom_text_repel(max.overlaps = 10, size = 3) +
        {if(ann)annotate("text", label= paste0("R = ", round(with(df_input,
                cor.test(total_deaths_per_million, get(x_metric), method = "pearson"))$estimate, 2),", p = ", 
                round(with(df_input,cor.test(total_deaths_per_million, get(x_metric), method = "pearson"))$p.value, 3),
                ", n = ", nrow(df_input)),x = min_x + (0.10 * (max_x - min_x)), y = max_y, color = "red", size = 3)} +
        labs(color = "Continent") +
        ylab("Total deaths per million") +
        xlab(x_label) +
        {if(Switch)ggtitle(paste0("Total deaths per million vs ", x_label, " as of ", date))} +
        guides(color = guide_legend(override.aes = list(size = 7))) +
        theme_minimal()+
        theme(axis.title = element_text(size = 13, family="serif"),
              axis.text = element_text(size = 11, family="serif"),
              legend.title = element_text(size = 13,family="serif"),
              legend.text = element_text(size = 10,family="serif"),
              plot.title = element_text(size = 14, family="serif"),
              plot.title.position = 'plot'
              )
  return(plot)
}

```



# **European Countries - 2020-12-30**

### Maps
```{r,echo=FALSE, message=FALSE, warning=FALSE}
ggarrange(
country_map(df_europe_20201230,'2020-12-30',"stringency","Europe","Stringency",lally_countries_map,'no'),
country_map(df_europe_20201230,'2020-12-30',"sub_stringency","Europe","Sub\nStringency",lally_countries_map,'no'), 
country_map(df_europe_20201230,'2020-12-30',"sub_stringency_cubed","Europe","Sub\nStringency^3",lally_countries_map,'no'),
country_map(df_europe_20201230,'2020-12-30',"lockdowns_cubed","Europe","Lockdowns^3",lally_countries_map,'no'),
nrow = 2, ncol = 2) %>% 
annotate_figure(map, top = text_grob(paste0("Europe Map ", "2020-12-30"), color = "black", 
          face = "bold", size = 16))

```

### Correlations - total covid deaths per million
```{r,echo=FALSE, message=FALSE, warning=FALSE}
corr_table(df_europe_20201230)
#corr_plot(df_europe_20201230)
corr_matrix(df_europe_20201230)
```

### Scatter plots - total covid deaths per million vs. stringency / sub_stringency cubed
```{r,echo=FALSE, message=FALSE, warning=FALSE}
ggarrange(
 country_scatter(df_europe_20201230,"stringency","Stringency",'no') + rremove("ylab"),
country_scatter(df_europe_20201230,"sub_stringency_cubed","Sub-stringency^3",'no') + rremove("ylab"), ncol = 2, legend = F) %>% 
  annotate_figure(map, top = text_grob(paste0("Total Deaths per Million vs. Stringency " , "2020-12-30"), color = "black", 
          face = "bold", size = 14), left = text_grob("Total Deaths per Million", rot = 90, vjust = 1))
```

### Regression
##### Replication of regression analysis from [Lally (2022)](https://link.springer.com/content/pdf/10.1007/s40592-021-00148-y.pdf) paper. 
```{r,echo=FALSE, message=FALSE, warning=FALSE}
mlr_table_si(df_europe_20201230)
mlr_table_res(df_europe_20201230)
mlr_table_res_cubed(df_europe_20201230)
mlr_table_lock(df_europe_20201230)
mlr_table_lock(df_europe_20201230)
```

# **European Countries - 2021-12-30**

### Maps
```{r,echo=FALSE, message=FALSE, warning=FALSE}
ggarrange(
country_map(df_europe_20211230,'2021-12-30',"stringency","Europe","Stringency",lally_countries_map,'no'),
country_map(df_europe_20211230,'2021-12-30',"sub_stringency","Europe","Sub\nStringency",lally_countries_map,'no'), 
country_map(df_europe_20211230,'2021-12-30',"sub_stringency_cubed","Europe","Sub\nStringency^3",lally_countries_map,'no'),
country_map(df_europe_20211230,'2021-12-30',"lockdowns_cubed","Europe","Lockdowns^3",lally_countries_map,'no'),
nrow = 2, ncol = 2) %>% 
annotate_figure(map, top = text_grob(paste0("Europe Map ", '2021-12-30'), color = "black", 
          face = "bold", size = 16))
```

### Correlations
```{r,echo=FALSE, message=FALSE, warning=FALSE}
corr_table(df_europe_20211230)
#corr_plot(df_europe_20211230)
corr_matrix(df_europe_20211230)
```

### Scatter plots - total covid deaths per million vs. stringency / sub_stringency cubed
```{r,echo=FALSE, message=FALSE, warning=FALSE}
ggarrange(
 country_scatter(df_europe_20211230,"stringency","Stringency",'no') + rremove("ylab"),
country_scatter(df_europe_20211230,"sub_stringency_cubed","Sub-stringency^3",'no') + rremove("ylab"), ncol = 2, legend = F) %>% 
  annotate_figure(map, top = text_grob(paste0("Total Deaths per Million vs. Stringency " , "2021-12-30"), color = "black", 
          face = "bold", size = 14), left = text_grob("Total Deaths per Million", rot = 90, vjust = 1))

```

### Regression
```{r,echo=FALSE, message=FALSE, warning=FALSE}
mlr_print(create_model_full(df_europe_20211230))
mlr_vif(create_model_full(df_europe_20211230))

mlr_print(create_model_stringency(df_europe_20211230))
mlr_vif(create_model_stringency(df_europe_20211230))

mlr_print(create_model_sub_stringency(df_europe_20211230))
mlr_vif(create_model_sub_stringency(df_europe_20211230))

mlr_print(create_model_sub_stringency_cubed(df_europe_20211230))
mlr_vif(create_model_sub_stringency_cubed(df_europe_20211230))

mlr_print(create_model_lockdowns(df_europe_20211230))
mlr_vif(create_model_lockdowns(df_europe_20211230))

mlr_print(create_model_lockdowns_cubed(df_europe_20211230))
mlr_vif(create_model_lockdowns_cubed(df_europe_20211230))

```

# **European Countries - 2022-05-20**
### Maps
```{r,echo=FALSE, message=FALSE, warning=FALSE}
ggarrange(
country_map(df_europe_20220520,'2022-05-20',"stringency","Europe","Stringency",lally_countries_map,'no'),
country_map(df_europe_20220520,'2022-05-20',"sub_stringency","Europe","Sub\nStringency",lally_countries_map,'no'), 
country_map(df_europe_20220520,'2022-05-20',"sub_stringency_cubed","Europe","Sub\nStringency^3",lally_countries_map,'no'),
country_map(df_europe_20220520,'2022-05-20',"lockdowns_cubed","Europe","Lockdowns^3",lally_countries_map,'no'),
nrow = 2, ncol = 2) %>% 
annotate_figure(map, top = text_grob(paste0("Europe Map ", '2022-05-20'), color = "black", 
          face = "bold", size = 16))

```

### Correlations
```{r,echo=FALSE, message=FALSE, warning=FALSE}
corr_table(df_europe_20220520)
#corr_plot(df_europe_20220520)
corr_matrix(df_europe_20220520)
```

### Scatter plots - total deaths per million vs. stringency / total restrictions cubed
```{r,echo=FALSE, message=FALSE, warning=FALSE}

ggarrange(
 country_scatter(df_europe_20220520,"stringency","Stringency",'no') + rremove("ylab"),
country_scatter(df_europe_20220520,"sub_stringency_cubed","Sub-stringency^3",'no') + rremove("ylab"), ncol = 2, legend = F) %>% 
  annotate_figure(map, top = text_grob(paste0("Total Deaths per Million vs. Stringency " , "2022-05-20"), color = "black", 
          face = "bold", size = 14), left = text_grob("Total Deaths per Million", rot = 90, vjust = 1))
```

### Regression
```{r,echo=FALSE, message=FALSE, warning=FALSE}
mlr_print(create_model_full(df_europe_20220520))
mlr_vif(create_model_full(df_europe_20220520))

mlr_print(create_model_stringency(df_europe_20220520))
mlr_vif(create_model_stringency(df_europe_20220520))

mlr_print(create_model_sub_stringency(df_europe_20220520))
mlr_vif(create_model_sub_stringency(df_europe_20220520))

mlr_print(create_model_sub_stringency_cubed(df_europe_20220520))
mlr_vif(create_model_sub_stringency_cubed(df_europe_20220520))

mlr_print(create_model_lockdowns(df_europe_20220520))
mlr_vif(create_model_lockdowns(df_europe_20220520))

mlr_print(create_model_lockdowns_cubed(df_europe_20220520))
mlr_vif(create_model_lockdowns_cubed(df_europe_20220520))
```


#### Countries in the analysis
```{r,echo=FALSE, message=FALSE, warning=FALSE}
# List of countries in analysis
lally_countries %>% as_tibble() %>% 
  rename(Countries = value) %>%  
  kbl(caption = paste0("Countries in analysis (n = ", length(lally_countries),")")) %>% 
  kable_classic(full_width = F) 
# Number of countries in analysis 
```


### Data sources:
* [Our World in Data](https://covid.ourworldindata.org/data/owid-covid-data.csv)
* [The Oxford COVID-19 Government Response Tracker](https://raw.githubusercontent.com/OxCGRT/covid-policy-tracker/master/data/OxCGRT_latest.csv)
* [Average latitude & longitude for Countries](https://raw.githubusercontent.com/albertyw/avenews/master/old/data/average-latitude-longitude-countries.csv)
* [Obesity Data](https://github.com/camminady/overweight)

```{r,echo=FALSE, message=FALSE, warning=FALSE}
#rm(list=setdiff(ls()," "))
```
